[QOL-8518] split Paster commands from Click
